requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.34
psycopg[binary]==3.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
pandas==2.2.2
openpyxl==3.1.5
numpy==1.26.4
pgvector==0.3.4
python-dateutil==2.9.0.post0
pytest==8.3.3
